requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
social-auth-app-django==5.7.0
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==12.1.1
django-storages==1.14.4
boto3==1.35.0
django-filter==24.3
celery==5.4.0
redis==5.2.0
gunicorn==23.0.0
whitenoise==6.8.2